[update] Add function.

+++ src/orig/fuga.c
@@ -4,7 +4,7 @@
 #include "hoge.h"
 
 int32_t incDecMany(int32_t val, size_t times) {
-    for (int i = 0; i < times; ++i) {
+    for (size_t i = 0; i < times; ++i) {
         val = incDec(val);
     }
     return val;
